backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-dotenv==1.0.0
httpx==0.24.1
openai==0.28.0
google-auth==2.22.0
google-api-python-client==2.97.0
azure-cosmos==4.5.1
python-multipart==0.0.6
jinja2==3.1.2
pdfkit==1.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
PyPDF2==3.0.1
python-docx==0.8.11
pandas==2.0.3
openpyxl==3.1.2
aiohttp==3.8.5
